search-engine/requirements.txt
python-dotenv
nltk            
tqdm           
